requirements.txt
eel==0.18.1
Jinja2==3.1.4
pillow==11.0.0
pyinstaller==6.11.1
pywebview==5.3.2
watchdog==6.0.0